split graph and log functions in util.ts
