+++ src/core/internal.ts
@@ -13,4 +13,5 @@ export * from "./reactor"
 export * from "./bank"
 export * from "./event"
 export * from "./cli"
+export * from "./graph"
